Changed name of the solver file
